# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RingiFlow API
  description: ワークフロー管理システム RingiFlow の BFF API
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
paths:
...
  /api/v1/workflows:
    get:
      tags:
      - workflows
      summary: GET /api/v1/workflows
      description: |-
        自分のワークフロー一覧を取得する

        ## 処理フロー

        1. セッションから `tenant_id`, `user_id` を取得
        2. Core Service の `GET /internal/workflows` を呼び出し
        3. レスポンスを返す
      operationId: list_my_workflows
      responses:
        '200':
          description: 自分のワークフロー一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowSummaryData'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    post:
      tags:
      - workflows
      summary: POST /api/v1/workflows
      description: |-
        ワークフローを作成する（下書き）

        ## 処理フロー

        1. セッションから `tenant_id`, `user_id` を取得
        2. Core Service の `POST /internal/workflows` を呼び出し
        3. レスポンスを返す
      operationId: create_workflow
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
        required: true
      responses:
        '201':
          description: ワークフロー作成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 定義が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
...
  /api/v1/workflows/{display_number}/comments:
    get:
      tags:
      - workflows
      summary: GET /api/v1/workflows/{display_number}/comments
      description: |-
        ワークフローのコメント一覧を取得する

        ## 処理フロー

        1. セッションから `tenant_id` を取得
        2. Core Service の `GET /internal/workflows/by-display-number/{display_number}/comments` を呼び出し
        3. 200 OK + コメント一覧を返す
      operationId: list_comments
      parameters:
      - name: display_number
        in: path
        description: ワークフロー表示番号
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: コメント一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowCommentData'
        '404':
          description: ワークフローが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    post:
      tags:
      - workflows
      summary: POST /api/v1/workflows/{display_number}/comments
      description: |-
        ワークフローにコメントを投稿する

        ## 処理フロー

        1. セッションから `tenant_id`, `user_id` を取得
        2. Core Service の `POST /internal/workflows/by-display-number/{display_number}/comments` を呼び出し
        3. 201 Created + コメントを返す
      operationId: post_comment
      parameters:
      - name: display_number
        in: path
        description: ワークフロー表示番号
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommentRequest'
        required: true
      responses:
        '201':
          description: コメント投稿成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowCommentData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: ワークフローが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/workflows/{display_number}/resubmit:
    post:
      tags:
      - workflows
      summary: POST /api/v1/workflows/{display_number}/resubmit
      description: |-
        ワークフローを再申請する

        ## 処理フロー

        1. セッションから `tenant_id`, `user_id` を取得
        2. Core Service の `POST /internal/workflows/by-display-number/{dn}/resubmit` を呼び出し
        3. 200 OK + 更新されたワークフローを返す
      operationId: resubmit_workflow
      parameters:
      - name: display_number
        in: path
        description: ワークフロー表示番号
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResubmitWorkflowRequest'
        required: true
      responses:
        '200':
          description: 再申請成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: ワークフローが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: 競合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/workflows/{display_number}/steps/{step_display_number}/approve:
    post:
      tags:
      - workflows
      summary: POST /api/v1/workflows/{display_number}/steps/{step_display_number}/approve
      description: |-
        ワークフローステップを承認する

        ## 処理フロー

        1. セッションから `tenant_id`, `user_id` を取得
        2. Core Service の `POST /internal/workflows/by-display-number/{dn}/steps/by-display-number/{step_dn}/approve` を呼び出し
        3. 200 OK + 更新されたワークフローを返す
      operationId: approve_step
      parameters:
      - name: display_number
        in: path
        description: ワークフローの表示用連番
        required: true
        schema:
          type: integer
          format: int64
      - name: step_display_number
        in: path
        description: ステップの表示用連番
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRejectRequest'
        required: true
      responses:
        '200':
          description: 承認成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: ステップが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: 競合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/workflows/{display_number}/steps/{step_display_number}/reject:
    post:
      tags:
      - workflows
      summary: POST /api/v1/workflows/{display_number}/steps/{step_display_number}/reject
      description: |-
        ワークフローステップを却下する

        ## 処理フロー

        1. セッションから `tenant_id`, `user_id` を取得
        2. Core Service の `POST /internal/workflows/by-display-number/{dn}/steps/by-display-number/{step_dn}/reject` を呼び出し
        3. 200 OK + 更新されたワークフローを返す
      operationId: reject_step
      parameters:
      - name: display_number
        in: path
        description: ワークフローの表示用連番
        required: true
        schema:
          type: integer
          format: int64
      - name: step_display_number
        in: path
        description: ステップの表示用連番
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRejectRequest'
        required: true
      responses:
        '200':
          description: 却下成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: ステップが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: 競合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/workflows/{display_number}/steps/{step_display_number}/request-changes:
    post:
      tags:
      - workflows
      summary: POST /api/v1/workflows/{display_number}/steps/{step_display_number}/request-changes
      description: |-
        ワークフローステップを差し戻す

        ## 処理フロー

        1. セッションから `tenant_id`, `user_id` を取得
        2. Core Service の `POST /internal/workflows/by-display-number/{dn}/steps/by-display-number/{step_dn}/request-changes` を呼び出し
        3. 200 OK + 更新されたワークフローを返す
      operationId: request_changes_step
      parameters:
      - name: display_number
        in: path
        description: ワークフローの表示用連番
        required: true
        schema:
          type: integer
          format: int64
      - name: step_display_number
        in: path
        description: ステップの表示用連番
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRejectRequest'
        required: true
      responses:
        '200':
          description: 差し戻し成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: 権限なし
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: ステップが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: 競合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/workflows/{display_number}/submit:
    post:
      tags:
      - workflows
      summary: POST /api/v1/workflows/{display_number}/submit
      description: |-
        ワークフローを申請する

        ## 処理フロー

        1. セッションから `tenant_id` を取得
        2. Core Service の `POST /internal/workflows/by-display-number/{display_number}/submit` を呼び出し
        3. レスポンスを返す
      operationId: submit_workflow
      parameters:
      - name: display_number
        in: path
        description: ワークフロー表示番号
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitWorkflowRequest'
        required: true
      responses:
        '200':
          description: 申請成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: ワークフローが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
...
components:
  schemas:
    ApproveRejectRequest:
      type: object
      description: ステップ承認/却下リクエスト（BFF 公開 API）
      required:
      - version
      properties:
        version:
          type: integer
          format: int32
          description: 楽観的ロック用バージョン
        comment:
          type:
          - string
          - 'null'
          description: コメント（任意）
...
    CreateWorkflowRequest:
      type: object
      description: ワークフロー作成リクエスト（BFF 公開 API）
      required:
      - definition_id
      - title
      - form_data
      properties:
        definition_id:
          type: string
          format: uuid
          description: ワークフロー定義 ID
        title:
          type: string
          description: ワークフロータイトル
        form_data:
          description: フォームデータ
...
    PostCommentRequest:
      type: object
      description: コメント投稿リクエスト（BFF 公開 API）
      required:
      - body
      properties:
        body:
          type: string
          description: コメント本文（1〜2000 文字）
    ProblemDetails:
      type: object
      description: |-
        エラーレスポンス（RFC 9457 Problem Details）

        すべてのサービスで統一されたエラーレスポンス形式。
        `type` フィールドは URI で問題の種類を識別する。
      required:
      - type
      - title
      - status
      - detail
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 0
        detail:
          type: string
...
    ResubmitWorkflowRequest:
      type: object
      description: ワークフロー再申請リクエスト（BFF 公開 API）
      required:
      - form_data
      - approvers
      - version
      properties:
        form_data:
          description: 更新後のフォームデータ
        approvers:
          type: array
          items:
            $ref: '#/components/schemas/StepApproverRequest'
          description: 各承認ステップの承認者リスト
        version:
          type: integer
          format: int32
          description: 楽観的ロック用バージョン
...
    StepApproverRequest:
      type: object
      description: ステップ承認者リクエスト（BFF 公開 API）
      required:
      - step_id
      - assigned_to
      properties:
        step_id:
          type: string
          description: 定義 JSON のステップ ID
        assigned_to:
          type: string
          format: uuid
          description: 承認者のユーザー ID
    SubmitWorkflowRequest:
      type: object
      description: ワークフロー申請リクエスト（BFF 公開 API）
      required:
      - approvers
      properties:
        approvers:
          type: array
          items:
            $ref: '#/components/schemas/StepApproverRequest'
          description: 各承認ステップの承認者リスト
...
    UserRefData:
      type: object
      description: ユーザー参照データ（フロントエンドへの Serialize 用）
      required:
      - id
      - name
      properties:
        id:
          type: string
        name:
          type: string
...
    WorkflowCommentData:
      type: object
      description: ワークフローコメントデータ
      required:
      - id
      - posted_by
      - body
      - created_at
      properties:
        id:
          type: string
        posted_by:
          $ref: '#/components/schemas/UserRefData'
        body:
          type: string
        created_at:
          type: string
    WorkflowData:
      type: object
      description: |-
        ワークフロー詳細用データ（ステップ付き）

        詳細 API およびコマンド系 API のレスポンスで使用。
      required:
      - id
      - display_id
      - display_number
      - title
      - definition_id
      - status
      - version
      - form_data
      - initiated_by
      - steps
      - created_at
      - updated_at
      properties:
        id:
          type: string
        display_id:
          type: string
        display_number:
          type: integer
          format: int64
        title:
          type: string
        definition_id:
          type: string
        status:
          type: string
        version:
          type: integer
          format: int32
        form_data: {}
        initiated_by:
          $ref: '#/components/schemas/UserRefData'
        current_step_id:
          type:
          - string
          - 'null'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStepData'
        submitted_at:
          type:
          - string
          - 'null'
        completed_at:
          type:
          - string
          - 'null'
        created_at:
          type: string
        updated_at:
          type: string
...
    WorkflowStepData:
      type: object
      description: ワークフローステップデータ
      required:
      - id
      - display_id
      - display_number
      - step_id
      - step_name
      - step_type
      - status
      - version
      - created_at
      - updated_at
      properties:
        id:
          type: string
        display_id:
          type: string
        display_number:
          type: integer
          format: int64
        step_id:
          type: string
        step_name:
          type: string
        step_type:
          type: string
        status:
          type: string
        version:
          type: integer
          format: int32
        assigned_to:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/UserRefData'
        decision:
          type:
          - string
          - 'null'
        comment:
          type:
          - string
          - 'null'
        due_date:
          type:
          - string
          - 'null'
        started_at:
          type:
          - string
          - 'null'
        completed_at:
          type:
          - string
          - 'null'
        created_at:
          type: string
        updated_at:
          type: string
    WorkflowSummaryData:
      type: object
      description: |-
        ワークフロー一覧用データ（ステップなし）

        一覧 API のレスポンスで使用。`steps` フィールドを含まない。
      required:
      - id
      - display_id
      - display_number
      - title
      - definition_id
      - status
      - version
      - form_data
      - initiated_by
      - created_at
      - updated_at
      properties:
        id:
          type: string
        display_id:
          type: string
        display_number:
          type: integer
          format: int64
        title:
          type: string
        definition_id:
          type: string
        status:
          type: string
        version:
          type: integer
          format: int32
        form_data: {}
        initiated_by:
          $ref: '#/components/schemas/UserRefData'
        current_step_id:
          type:
          - string
          - 'null'
        submitted_at:
          type:
          - string
          - 'null'
        completed_at:
          type:
          - string
          - 'null'
        created_at:
          type: string
        updated_at:
          type: string
  securitySchemes:
    session_auth:
      type: apiKey
      in: cookie
      name: session_id
...